LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerlock/pointerevent_pointerlock_supercedes_capture-expected.txt

Pointer Events pointer lock test

Follow the test instructions with mouse. If you don't have the device skip it.

Test Description: This test checks that we do not set the pointer capture when any element in the page gets a pointer lock.
Press left button down on the green rectangle and hold it.
Move the mouse inside the green rectangle.
Test passes if the pointer capture is not set when the green rectangle gets the pointer lock.


FAIL mouse no pointercapture while pointerlock assert_greater_than: More than 1 pointermove has been received after unlocked expected a number greater than 1 but got 0
